+++ emspay.php
@@ -4,7 +4,7 @@
  * Plugin Name: EMS Online
  * Plugin URI: https://emspay.nl/
  * Description: EMS Pay WooCommerce plugin
- * Version: 1.1.4
+ * Version: 1.1.5
  * Author: Ginger Payments
  * Author URI: https://www.gingerpayments.com/
  * License: The MIT License (MIT)
@@ -69,10 +69,33 @@ function woocommerce_emspay_init()
         return $methods;
     }
 
-    add_filter('woocommerce_payment_gateways', 'woocommerce_add_emspay');
-    add_action('woocommerce_api_callback', array(new woocommerce_emspay(), 'handle_callback'));
+    /**
+     * Check if Klarna payment method is limited to specific set of IPs.
+     *
+     * @param $gateways
+     * @return mixed
+     */
+    function ginger_klarna_enabled_ip($gateways)
+    {
+        $settings = get_option('woocommerce_emspay_settings');
+        $ems_klarna_ip_list = $settings['debug_klarna_ip'];
 
-    function register_shipped_order_status()
+        if (strlen($ems_klarna_ip_list) > 0) {
+            $ip_whitelist = array_map('trim', explode(",", $ems_klarna_ip_list));
+
+            if (!in_array(WC_Geolocation::get_ip_address(), $ip_whitelist)) {
+                unset($gateways['emspay_klarna-pay-later']);
+            }
+        }
+
+        return $gateways;
+    }
+
+    add_filter('woocommerce_available_payment_gateways', 'ginger_klarna_enabled_ip');
+    add_filter('woocommerce_payment_gateways', 'woocommerce_add_emspay');
+    add_action('woocommerce_api_callback', array(new woocommerce_emspay(), 'ginger_handle_callback'));
+
+    function ginger_register_shipped_order_status()
     {
         register_post_status('wc-shipped', array(
             'label' => 'Shipped',
@@ -84,13 +107,13 @@ function woocommerce_emspay_init()
         ));
     }
 
-    add_action('init', 'register_shipped_order_status');
+    add_action('init', 'ginger_register_shipped_order_status');
 
     /**
      * @param array $order_statuses
      * @return array
      */
-    function add_shipped_to_order_statuses(array $order_statuses)
+    function ginger_add_shipped_to_order_statuses(array $order_statuses)
     {
         $new_order_statuses = array();
         foreach ($order_statuses as $key => $status) {
@@ -102,25 +125,24 @@ function woocommerce_emspay_init()
         return $new_order_statuses;
     }
 
-    add_filter('wc_order_statuses', 'add_shipped_to_order_statuses');
-
-    add_action('woocommerce_order_status_shipped', 'ship_an_order', 10, 2);
-    add_action('woocommerce_refund_created', 'refund_an_order', 10, 2);
-    add_action('woocommerce_order_item_add_action_buttons', 'add_refund_description');
+    add_filter('wc_order_statuses', 'ginger_add_shipped_to_order_statuses');
+    add_action('woocommerce_order_status_shipped', 'ginger_ship_an_order', 10, 2);
+    add_action('woocommerce_refund_created', 'ginger_refund_an_order', 10, 2);
+    add_action('woocommerce_order_item_add_action_buttons', 'ginger_add_refund_description');
 
     load_plugin_textdomain(WC_Emspay_Helper::DOMAIN, false, basename(dirname(__FILE__)).'/languages');
 
 	/**
-	 * Function refund_an_order - refund EMS order
+	 * Function ginger_refund_an_order - refund EMS order
 	 *
 	 * @param $refund_id
 	 * @param $args
 	 */
-    function refund_an_order($refund_id, $args) {
+    function ginger_refund_an_order($refund_id, $args) {
 
 		$ems_order_id = get_post_meta($args['order_id'], 'ems_order_id', true);
 		$order = wc_get_order($args['order_id']);
-		$ginger = get_ginger_client($order);
+		$ginger = ginger_get_client($order);
 		$emsOrder = $ginger->getOrder($ems_order_id);
 		$orderGateway = $order->get_payment_method();
 		
@@ -129,7 +151,7 @@ function woocommerce_emspay_init()
 		}
 		
 		$refund_data = [
-			'amount' => WC_Emspay_Helper::getAmountInCents($args['amount']), 
+			'amount' => WC_Emspay_Helper::gingerGetAmountInCents($args['amount']),
 			'description' => 'OrderID: #' . $args['order_id'] . ', Reason: ' . $args['reason']
 		];
 	
@@ -137,7 +159,7 @@ function woocommerce_emspay_init()
 			if(!isset($emsOrder['transactions']['flags']['has-captures'])) {
 				throw new Exception( 'Refunds only possible when captured' );
 			};
-			$refund_data['order_lines'] = WC_Emspay_Helper::getOrderLines($order);
+			$refund_data['order_lines'] = WC_Emspay_Helper::gingerGetOrderLines($order);
 		}
 
 		update_post_meta($args['order_id'], 'refund_id', $refund_id);
@@ -156,25 +178,25 @@ function woocommerce_emspay_init()
 	}
 	
 	/**
-	 * Function add_refund_description
+	 * Function ginger_add_refund_description
 	 *
 	 * @param $order
 	 */
-	function add_refund_description($order) {
-		echo "<p style='color: red; ' class='description'>" . __( "Please beware that for EMS transactions the refunds will process directly to the gateway!", "emspay") . "</p>";
+	function ginger_add_refund_description($order) {
+		echo "<p style='color: red; ' class='description'>" . esc_html__( "Please beware that for EMS transactions the refunds will process directly to the gateway!", "emspay") . "</p>";
 	}
 
 	/**
-	 * Function ship_an_order - Support for Klarna and Afterpay order shipped state
+	 * Function ginger_ship_an_order - Support for Klarna and Afterpay order shipped state
 	 *
 	 * @param $order_id
 	 * @param $order
 	 */
-    function ship_an_order($order_id, $order)
+    function ginger_ship_an_order($order_id, $order)
     {
         if ($order && $order->get_status() == 'shipped' && in_array($order->get_payment_method(), array('emspay_klarna-pay-later', 'emspay_afterpay'))) {
 
-			$ginger = get_ginger_client($order);
+			$ginger = ginger_get_client($order);
 
             try {
                 $id = get_post_meta($order_id, 'ems_order_id', true);
@@ -188,12 +210,12 @@ function woocommerce_emspay_init()
     }
 
 	/**
-	 * Function get_ginger_client
+	 * Function ginger_get_client
 	 *
 	 * @param $order
 	 * @return \Ginger\ApiClient
 	 */
-    function get_ginger_client($order) {
+    function ginger_get_client($order) {
 		$settings = get_option('woocommerce_emspay_settings');
 		$apiKey = $settings['api_key'];
 
@@ -217,7 +239,7 @@ function woocommerce_emspay_init()
 				$apiKey,
 				($settings['bundle_cacert'] == 'yes') ?
 					[
-						CURLOPT_CAINFO => WC_Emspay_Helper::getCaCertPath()
+						CURLOPT_CAINFO => WC_Emspay_Helper::gingerGetCaCertPath()
 					] : []
 			);
 		} catch (Exception $exception) {
@@ -234,9 +256,9 @@ function woocommerce_emspay_init()
      * @param WC_Order $order
      * @return string
      */
-    function emspay_order_received_text($text, $order)
+    function ginger_order_received_text($text, $order)
     {
-        return WC_Emspay_Helper::getOrderDescription($order->get_id());
+        return WC_Emspay_Helper::gingerGetOrderDescription($order->get_id());
     }
 
     /**
@@ -245,7 +267,7 @@ function woocommerce_emspay_init()
      * @param array $gateways
      * @return mixed
      */
-    function afterpay_filter_gateway($gateways)
+    function ginger_afterpay_filter_gateways($gateways)
     {
         if ( ! is_checkout() ) {
             return $gateways;
@@ -279,7 +301,7 @@ function woocommerce_emspay_init()
      * @param array $gateways
      * @return mixed
      */
-    function klarna_filter_gateway($gateways)
+    function ginger_klarna_filter_gateway($gateways)
     {
         if ( ! is_checkout() ) {
             return $gateways;
@@ -305,7 +327,7 @@ function woocommerce_emspay_init()
      * @param $gateways
      * @return bool
      */
-    function filter_gateway_by_currency($gateways) {
+    function ginger_filter_gateway_by_currency($gateways) {
         if ( ! is_checkout() ) {
             return $gateways;
         }
@@ -327,8 +349,8 @@ function woocommerce_emspay_init()
         return $gateways;
     }
 
-    add_filter('woocommerce_available_payment_gateways', 'afterpay_filter_gateway', 10);
-    add_filter('woocommerce_available_payment_gateways', 'klarna_filter_gateway', 10);
-    add_filter('woocommerce_available_payment_gateways', 'filter_gateway_by_currency', 10);
-    add_filter('woocommerce_thankyou_order_received_text', 'emspay_order_received_text', 10, 2);
+    add_filter('woocommerce_available_payment_gateways', 'ginger_afterpay_filter_gateway', 10);
+    add_filter('woocommerce_available_payment_gateways', 'ginger_klarna_filter_gateway', 10);
+    add_filter('woocommerce_available_payment_gateways', 'ginger_filter_gateway_by_currency', 10);
+    add_filter('woocommerce_thankyou_order_received_text', 'ginger_order_received_text', 10, 2);
 }